+++ src/sdk.js
@@ -164,6 +164,13 @@ const endpointDefinitions = [
   },
   {
     apiName: 'api',
+    path: 'current_user/delete',
+    internal: false,
+    method: 'post',
+    interceptors: [new TransitResponse(), new TransitRequest()],
+  },
+  {
+    apiName: 'api',
     path: 'current_user/verify_email',
     internal: false,
     method: 'post',